WebStory/requirements.txt
autocorrect==2.6.1
Flask==2.2.3
gensim==3.6.0
ipython==7.33.0
nltk==3.8.1
openai==0.27.4
Pillow==9.5.0
Requests==2.28.2
